line-level-data-collection/requirements.txt
loguru==0.5.3
psycopg2-binary==2.9.5
boto3==1.20.32
dataclasses-json==0.5.7